// packs/data/gametests/src/waila/types/LookAtObjectMetadataInterface.ts
export type ItemStackWithSlot = {
	item: ItemStack;
	slot: number;
};

/**
 * Rendering data specifically for blocks
 */
export interface BlockRenderDataInterface {
	/**
	 * Renderer string that will be used to display the tools that can be used to efficiently break the block.
	 */
	toolIcons: string;
	/**
	 * The block states of the block being looked at.
	 */
	blockStates?: string;
	/**
	 * Content or contents within the block.
	 * This is used for blocks that can contain items, such as chests or barrels.
	 */
	inventory?: ItemStackWithSlot[];
	/** Number of additional item stacks that could not be mirrored into the player inventory projection. */
	inventoryOverflow?: number;
	/**
	 * R4.8: 额外详细信息（堆肥/蜂蜜/蜡烛/红石/燃料等），蹲下时显示。
	 * 每行是一个 RawMessage 数组（支持 translate 中文）
	 */
	extraInfo?: RawMessage[][];
}